Lighting controller code update
This version of the lighting controller has a very noticable delay in
the animation cycle while it reads the dip switch (even without the
settle delay in the dip switch function). CircuitPython does not
natively support any kind of threading or async. Researching this
but might convert to native Arduino if that can handle it better.